requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
kafka-python==2.0.2
flask==2.3.3
flask-cors==4.0.0
